+++ src/Day02.kt
@@ -1,4 +1,4 @@
-typealias Report = List<Long>
+typealias Report = List<Int>
 
 fun main() {
 
@@ -15,23 +15,23 @@ fun main() {
     }
 
     fun inputToReports(input: List<String>): List<Report> {
-        return input.map { line -> line.split(" ").map { it.toLong() } }
+        return input.map { line -> line.split(" ").map { it.toInt() } }
     }
 
-    fun part1(input: List<String>): Long {
+    fun part1(input: List<String>): Int {
         val reports = inputToReports(input)
-        return reports.count { it.isSafe() }.toLong()
+        return reports.count { it.isSafe() }
     }
 
-    fun part2(input: List<String>): Long {
+    fun part2(input: List<String>): Int {
         val reports = inputToReports(input)
-        return reports.count { it.isSafe(false) }.toLong()
+        return reports.count { it.isSafe(false) }
     }
 
     // test before attempt to solve
     val testInput = readInput("Day02_test")
-    check(part1(testInput) == 2L)
-    check(part2(testInput) == 4L)
+    check(part1(testInput) == 2)
+    check(part2(testInput) == 4)
 
     // solve with real input
     solve("Day02", ::part1, ::part2)
