+++ src/Day03.kt
@@ -3,20 +3,20 @@ fun main() {
     val instructionPattern = Regex("""mul\((\d+),(\d+)\)""")
     val deactivationPattern = Regex("""don't\(\).+?(do\(\)|$)""")
 
-    fun String.addMultiplications(): Long {
+    fun String.addMultiplications(): Int {
         return instructionPattern
             .findAll(this)
             .sumOf {
                 val (a, b) = it.destructured
-                a.toLong() * b.toLong()
+                a.toInt() * b.toInt()
             }
     }
 
-    fun part1(input: List<String>): Long {
+    fun part1(input: List<String>): Int {
         return input.joinToString().addMultiplications()
     }
 
-    fun part2(input: List<String>): Long {
+    fun part2(input: List<String>): Int {
         return input
             .joinToString()
             .replace(deactivationPattern, "")
@@ -25,8 +25,8 @@ fun main() {
 
     // test before attempt to solve
     val testInput = readInput("Day03_test")
-    check(part1(listOf(testInput.first())) == 161L)
-    check(part2(listOf(testInput.last())) == 48L)
+    check(part1(listOf(testInput.first())) == 161)
+    check(part2(listOf(testInput.last())) == 48)
 
     // solve with real input
     solve("Day03", ::part1, ::part2)
